+++ GoodNews.Client/src/App.js
@@ -4,8 +4,9 @@ import './App.css';
 import Login from "./Components/Login";
 import Register from "./Components/Register";
 import Header from "./Components/Header";
-import News from "./Components/News";
+// import News from "./Components/News";
 import ArticleDetails from "./Components/ArticleDetails";
+import NewsList from "./Components/NewsList";
 
 const App = () => {
     return (
@@ -13,8 +14,8 @@ const App = () => {
             <div className="App">
                 <Header/>
                 <div className='app-content-wrapper'>
-                    <Route exact path='/' component={News}/>
-                    <Route exact path='/news' component={News}/>
+                    <Route exact path='/' component={NewsList}/>
+                    <Route exact path='/news' component={NewsList}/>
                     <Route path='/news/:id' component={ArticleDetails}/>
                     <Route path='/login' component={Login}/>
                     <Route path='/register' component={Register}/>
