+++ GoodNews.Client/src/Components/News/ArticleCard/style.module.scss
@@ -1,3 +1,4 @@
 a {
   text-decoration: none;
+  color: black;
 }
