+++ GoodNews.Client/src/Components/NewsList/style.module.scss
@@ -1,0 +1,3 @@
+.newsContainer {
+  margin-top: 20px;
+}
